#24 Reorganize macro packages. Added version in title. Bumped minimum Fiji version needed to run plugin.
