+++ Maize/src/cimmyt/maize/ui/macros/canopy/Maize_CIMMYT_QuanitfyImageComponentsHSB_UBv8.ijm
