+++ Maize/src/plugins.config
@@ -6,7 +6,7 @@
 
 
 # Author: George El-Haddad
-# Version: 0.12
+# Version: 0.13
 # Date: 2015/11/05
 # Requires: ImageJ 1.49m
